know-1703: Added an index-js file for each lottie folder.
